services/predictor/pyproject.toml
[project]
name = "predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "risingwave-py>=0.0.2",
    "psycopg2-binary>=2.9.9",
    "great-expectations>=0.18.0",
    "numpy>=1.26.0",
    "ydata-profiling>=4.16.1",
    "loguru>=0.7.3",
    "mlflow>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
